LuLib 2.0

+++ include/LuConstants.h
@@ -1,15 +1,15 @@
 #ifndef LU_CONSTANTS_H
 #define LU_CONSTANTS_H
 
-#define PI            ((float)(3.14159265)) // PI or 180°
-#define TWO_PI        ((float)(6.28318530)) // PI * 2 or 360°
-#define HALF_PI       ((float)(1.57079632)) // PI / 2 or 90°
-#define QUARTER_PI    ((float)(0.78539816)) // PI / 4 or 45°
-#define DEG_TO_RAD    ((float)(0.01745329)) // PI / 180
-#define RAD_TO_DEG   ((float)(57.29577951)) // 180 / PI
-#define PYTHAGORAS    ((float)(1.41421356)) // sqrt(2) my favorite constant
-#define E             ((float)(2.71828182)) // Euler's number : inf sum of 1/(n!) with n starting at 0
-#define GOLDEN_RATIO  ((float)(1.61803398)) // the "perfect" ratio
-#define PHI           ((float)(1.61803398)) // another name for the golden ratio, used in math
+#define PI (3.14159265f)			// PI or 180 deg
+#define TWO_PI (6.28318530f)		// PI * 2 or 360 deg
+#define HALF_PI (1.57079632f)		// PI / 2 or 90 deg
+#define QUARTER_PI (0.78539816f)	// PI / 4 or 45 deg
+#define DEG_TO_RAD (0.01745329f)	// PI / 180
+#define RAD_TO_DEG (57.29577951f)	// 180 / PI
+#define PYTHAGORAS (1.41421356f)	// sqrt(2)
+#define E (2.71828182f)				// Euler's number
+#define GOLDEN_RATIO (1.61803398f)	// The "perfect" ratio
+#define PHI (1.61803398f)			// Another name for GOLDEN_RATIO
 
 #endif	// LU_CONSTANTS_H
